uint128 (#60)
* uint128

* comments

* add library

* tests

+++ src/libraries/ParamsBuilder.sol
@@ -47,8 +47,8 @@ library ParamsBuilder {
         // Build parameters
         params = new bytes[](paramsArraySize);
 
-        uint256 amount0 = currencyIsCurrency0 ? fullRangeParams.currencyAmount : fullRangeParams.tokenAmount;
-        uint256 amount1 = currencyIsCurrency0 ? fullRangeParams.tokenAmount : fullRangeParams.currencyAmount;
+        uint128 amount0 = currencyIsCurrency0 ? fullRangeParams.currencyAmount : fullRangeParams.tokenAmount;
+        uint128 amount1 = currencyIsCurrency0 ? fullRangeParams.tokenAmount : fullRangeParams.currencyAmount;
 
         // Set up mint
         params[0] = abi.encode(
